backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
requests>=2.28.0
docker>=6.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.5
python-multipart>=0.0.6
alembic>=1.11.1
sqlalchemy>=2.0.0
